fix: transformCliUtils was returning undefined in some cases
